improve deps (drop unnecessary Requires: python-module-setuptools)
That was a RedHatism. I'd say, in ALT, we prefer to rely on autoreqs
rather than explicit Requires unless they are insufficient; then, we
add only the necessary explicit Requires.

This dep must have been due to the use of pkg_resources from
setuptools, already covered by autoreqs (python*(pkg_resources)).

If this is true, dropping the explicit Requires is an improvement,
because it didn't give a precise idea what was used (from setuptools).
Now, we have more precise information for solving problems with
transitive deps (like an unwanted heavy dep on gcc).