Add support for the stacked backup
The current implementation doesn't ensure the order of
backup/restoration. There are use cases requiring the multiple
backups of a file with same name but different content. And for
these situations, FileStore is not predictable.

To ensure the order usual dict storage is replaced with
OrderedDict one. Also, all search operations are performed in
a reversed order.

Fixes: https://pagure.io/freeipa/issue/7820
Signed-off-by: Stanislav Levin <slev@altlinux.org>