spec: Fix `without_lint` build
Actually, `setuptools` is the direct build dependency of Python IPA.

For `with_lint` build `setuptools` is in build environment due to the
transitive dependency of `pylint`. For `without_lint` build there is
no such.

Thus, `setuptools` should be specified as BR for all build scenarios.
Thanks mike@ for reporting the issue.