Fix pytest4 errors (xmlrpc)
This migrates `setup`/`teardown` functions/methods to the
powerfull pytest fixtures:
```
The purpose of test fixtures is to provide a fixed baseline upon
which tests can reliably and repeatedly execute. pytest fixtures
offer dramatic improvements over the classic xUnit style of
setup/teardown functions:

 - fixtures have explicit names and are activated by declaring their
 use from test functions, modules, classes or whole projects.

 - fixtures are implemented in a modular manner, as each fixture name
 triggers a fixture function which can itself use other fixtures.

 - fixture management scales from simple unit to complex functional
 testing, allowing to parametrize fixtures and tests according to
 configuration and component options, or to re-use fixtures across
 function, class, module or whole test session scopes.
```

This a part of work to remove the mixing of test idioms in the
test suite.

Signed-off-by: Stanislav Levin <slev@altlinux.org>