strip-po: Add dependency on stamp-po
The targets specified as dependencies of another target are built
in parallel. Both `stamp-po` and `strip-po` do want the write
access for po and pot files. This becomes racy and leads to
occasional build errors.

Since `strip-po` do its operations on ready-for-use po and pot the
`stamp-po` target should be its direct dependency.

$(DOMAIN).pot is the dependency of `stamp-po` and can be omitted.

Fixes: https://pagure.io/freeipa/issue/8323