ipatests: Add support for remote ipaplatform
Previously, there was the assumption that platform's OS and
platform's OS version of tests' controller and tests' workers are
the same. But actually, for example, a controller can be Fedora,
while the worker RHEL. Workers in turn can differ too, for example,
worker1 - Debian, worker2 - Fedora.

For convenience a remote ipaplatform is fetched on demand into
controller and cached for the future usage.

Currently, only the JSON serializable objects of remote platform are
gathered as most usable, but there is no problem to extend the
corresponding classes.

Fixes: https://pagure.io/freeipa/issue/8620
Signed-off-by: Stanislav Levin <slev@altlinux.org>